异常查找

+++ src/main/java/com/tencent/wxcloudrun/model/RefurbishmentResult.java
@@ -7,10 +7,10 @@ import java.io.Serializable;
 @Data
 public class RefurbishmentResult implements Serializable {
 
-    private long completed_at;
-    private long created_at;
+    private Integer completed_at;
+    private Integer created_at;
     private String image;
-    private long processed_at;
+    private Integer processed_at;
     private Integer progress;
     private Integer return_type;
     private Integer state;
